keymgr: fix 'array subscript type' warning on NetBSD

+++ src/utils/keymgr/functions.c
@@ -32,10 +32,10 @@
 
 static time_t arg_timestamp(const char *arg)
 {
-	if (isdigit(arg[0]) && strlen(arg) < 12) {
+	if (isdigit((int)arg[0]) && strlen(arg) < 12) {
 		return atol(arg); // unixtime
 	}
-	if (isdigit(arg[0]) && strlen(arg) == 14) {
+	if (isdigit((int)arg[0]) && strlen(arg) == 14) {
 		struct tm tm = { 0 };
 		char *end = strptime(arg, "%Y%m%d%H%M%S", &tm);
 		if (end == NULL || *end != '\0') {
@@ -99,7 +99,7 @@ static bool genkeyargs(int argc, char *argv[], bool just_timing,
 	// parse args
 	for (int i = 0; i < argc; i++) {
 		if (!just_timing && strncasecmp(argv[i], "algorithm=", 10) == 0) {
-			if (isdigit(argv[i][10]) && atol(argv[i] + 10) < 256) {
+			if (isdigit((int)argv[i][10]) && atol(argv[i] + 10) < 256) {
 				*algorithm = atol(argv[i] + 10);
 				continue;
 			}
@@ -116,7 +116,7 @@ static bool genkeyargs(int argc, char *argv[], bool just_timing,
 				return false;
 			}
 		} else if (!just_timing && strncasecmp(argv[i], "ksk=", 4) == 0) {
-			switch (tolower(argv[i][4])) {
+			switch (tolower((int)argv[i][4])) {
 			case '1':
 			case 'y':
 			case 't':
@@ -424,7 +424,7 @@ static long is_uint32(const char *string)
 		return -1;
 	}
 	for (const char *p = string; *p != '\0'; p++) {
-		if (!isdigit(*p)) {
+		if (!isdigit((int)*p)) {
 			return -1;
 		}
 	}
@@ -435,7 +435,7 @@ static long is_uint32(const char *string)
 static bool is_hex(const char *string)
 {
 	for (const char *p = string; *p != '\0'; p++) {
-		if (!isxdigit(*p)) {
+		if (!isxdigit((int)*p)) {
 			return false;
 		}
 	}
